Created new class in discord_bot.py with all async functions moved out of main
